--- Assets/Render Texture/Autopilot.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9f371e682f2f7694796c35b2b51355c1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
