--- Assets/Render Texture/GhostPERRINN 424.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 72ead883c3ef5ff49bacd5e06260cd68
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
