--- Assets/Render Texture/SteeringScreen.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ddf33aec416ecc94285e7a9ffafa90d4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
