+++ Assets/RenderTexture/SteeringDisplay.renderTexture.meta
