+++ Assets/RenderTexture/SteeringScreen.cs.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: 99f341b71b410314b9f2d361260f98c6
+guid: 10968d37f1acb20469b51382049f6f79
+timeCreated: 1482715088
+licenseType: Pro
 MonoImporter:
-  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
